Making setDatapoint return unchanged state if the old and new displayValues match
This is based on the idea that any change in the dataValue of the
data point will be reflected in the display value as well, which
may prove to be an insufficient heuristic later, in which case
we would need to do a deep comparison of the dataValue object. That'll
be way worse for performance though, so making this work if we can
will make things faster.