Trying to stop the message stack from looping on top of itself
I'm not sure this'll work though because inside the for loop everything
is async. So it'll probably just execute the loop, set everything to
pending, and then once it starts firing async status changes due to
the database and network, it'll be in "ready" position again, so it'll
still loop over all remaining messages for every network response.