First pass at setting up routing for project page / dataset page
This has significant issues because it's using the hash router
to force S3 to support client-side routing. Ideally, these pages
should just be generated as static pages (with all their metadata)
so that they are SEO optimized and loading them doesn't touch the
database. That's supported in gatsby-hosted-gatsby, but not in S3
because we don't have a stateful server which can do incremental
static generation.